requirements.txt
openai==0.28.0
streamlit==1.27.2
python-dotenv
